requirements-new.txt
attrs==23.2.0
aws-cdk-lib==2.135.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
cattrs==23.2.3
constructs==10.3.0
demo-tools==0.0.3
exceptiongroup==1.2.0
importlib_resources==6.4.0
jsii==1.97.0
publication==0.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.11.0
zipp==3.18.1
